@@ -5,17 +5,17 @@ const http = require('http');
 app.use(express.static('dist'));
 app.use(express.json())
 
-let map = {}
+let map = new Map()
 
 app.post("/send", (req, res) => {
-    map[req.body.movie] = req.body
+    map.set(req.body.movie, req.body)
     res.json("OK")
 });
 
 
 app.post("/receive", (req, res) => {
-    res.json(map[req.body.movie])
+    res.json(map.get(req.body.movie))
 });
 
-http.createServer(app).listen(7799);
+http.createServer(app).listen(7799,"::0");
 
